Prepare for maven central release
