--- src/main/java/com/github/simplecqrsjava/core/queries/Query.java
@@ -1,10 +1,0 @@
-package com.github.simplecqrsjava.core.queries;
-
-/**
- * A query represents the retrieval of data from the system, this should contain all necessary to retrieve the response for this query.
- *
- * @param <R> - data that is expected to be provided as part of the query, for the query handling
- */
-@SuppressWarnings("unused")
-public interface Query<R> {
-}
