--- src/main/java/com/github/simplecqrsjava/core/shared/exceptions/UnregisteredHandlerException.java
@@ -1,8 +1,0 @@
-package com.github.simplecqrsjava.core.shared.exceptions;
-
-public class UnregisteredHandlerException extends RuntimeException {
-
-    public UnregisteredHandlerException(Class<?> classType) {
-        super(String.format("Handler for %s has not been registered", classType.getSimpleName()));
-    }
-    }
