--- src/test/java/com/github/simplecqrsjava/core/commands/fakes/FakeCommand.java
@@ -1,6 +1,0 @@
-package com.github.simplecqrsjava.core.commands.fakes;
-
-import com.github.simplecqrsjava.core.commands.Command;
-
-public class FakeCommand implements Command<Boolean> {
-}
