--- src/test/java/com/github/simplecqrsjava/core/events/fakes/FakeEvent.java
@@ -1,9 +1,0 @@
-package com.github.simplecqrsjava.core.events.fakes;
-
-import com.github.simplecqrsjava.core.events.Event;
-
-public class FakeEvent extends Event<Long> {
-    public FakeEvent(Long data) {
-        super(data);
-    }
-}
