--- src/test/java/com/github/simplecqrsjava/core/events/fakes/FakeEventHandler.java
@@ -1,10 +1,0 @@
-package com.github.simplecqrsjava.core.events.fakes;
-
-import com.github.simplecqrsjava.core.events.EventHandler;
-
-public class FakeEventHandler implements EventHandler<FakeEvent> {
-    @Override
-    public void handle(FakeEvent event) {
-        //Stubbed method
-    }
-}
