--- src/test/java/com/github/simplecqrsjava/core/queries/fakes/FakeQuery.java
@@ -1,6 +1,0 @@
-package com.github.simplecqrsjava.core.queries.fakes;
-
-import com.github.simplecqrsjava.core.queries.Query;
-
-public class FakeQuery implements Query<Boolean> {
-}
