--- src/test/java/com/github/simplecqrsjava/core/queries/fakes/FakeQueryHandler.java
@@ -1,10 +1,0 @@
-package com.github.simplecqrsjava.core.queries.fakes;
-
-import com.github.simplecqrsjava.core.queries.QueryHandler;
-
-public class FakeQueryHandler implements QueryHandler<Boolean, FakeQuery> {
-    @Override
-    public Boolean handle(FakeQuery query) {
-        return true;
-    }
-}
