+++ src/test/java/io/github/broot02/simplecqrsjava/core/commands/fakes/FakeCommand.java
@@ -1,0 +1,6 @@
+package io.github.broot02.simplecqrsjava.core.commands.fakes;
+
+import io.github.broot02.simplecqrsjava.core.commands.Command;
+
+public class FakeCommand implements Command<Boolean> {
+}
