+++ src/test/java/io/github/broot02/simplecqrsjava/core/commands/fakes/FakeCommandHandler.java
@@ -1,0 +1,10 @@
+package io.github.broot02.simplecqrsjava.core.commands.fakes;
+
+import io.github.broot02.simplecqrsjava.core.commands.CommandHandler;
+
+public class FakeCommandHandler implements CommandHandler<FakeCommand, Boolean> {
+    @Override
+    public Boolean handle(FakeCommand command) {
+        return true;
+    }
+}
