+++ src/test/java/io/github/broot02/simplecqrsjava/core/events/fakes/FakeEvent.java
@@ -1,0 +1,9 @@
+package io.github.broot02.simplecqrsjava.core.events.fakes;
+
+import io.github.broot02.simplecqrsjava.core.events.Event;
+
+public class FakeEvent extends Event<Long> {
+    public FakeEvent(Long data) {
+        super(data);
+    }
+}
