+++ src/test/java/io/github/broot02/simplecqrsjava/core/events/fakes/FakeEventHandler.java
@@ -1,0 +1,10 @@
+package io.github.broot02.simplecqrsjava.core.events.fakes;
+
+import io.github.broot02.simplecqrsjava.core.events.EventHandler;
+
+public class FakeEventHandler implements EventHandler<FakeEvent> {
+    @Override
+    public void handle(FakeEvent event) {
+        //Stubbed method
+    }
+}
