+++ src/test/java/io/github/broot02/simplecqrsjava/core/queries/fakes/FakeQuery.java
@@ -1,0 +1,6 @@
+package io.github.broot02.simplecqrsjava.core.queries.fakes;
+
+import io.github.broot02.simplecqrsjava.core.queries.Query;
+
+public class FakeQuery implements Query<Boolean> {
+}
