+++ src/test/java/io/github/broot02/simplecqrsjava/core/queries/fakes/FakeQueryHandler.java
@@ -1,0 +1,10 @@
+package io.github.broot02.simplecqrsjava.core.queries.fakes;
+
+import io.github.broot02.simplecqrsjava.core.queries.QueryHandler;
+
+public class FakeQueryHandler implements QueryHandler<Boolean, FakeQuery> {
+    @Override
+    public Boolean handle(FakeQuery query) {
+        return true;
+    }
+}
